Add basic poll rendering from localstorage
Use a path like `/polls/1/` to designate the poll id. Look up that ID in
localstorage.

For now, I've hardcoded a few initial polls. Next steps:
* Make the polls array editable by instructor
* Propagate changes to the polls array via websockets?

Currently, there is no notion of instructor/student in the application.
This is fine, but we will need to have some mechanism here for giving an
"instructor" view of the poll, which will look very different from the
students' view. The thing is, we're designing this application without a
database, at least to start. We can think creatively on how to solve
this. And of course, we can always end up using a database at the end of
the day, if necessary.