Convert levelWorker to a Promise
This promise is still synchronous
(https://stackoverflow.com/a/36588964/173630) but I'm not seeing too
much slowdown when re-rendering here, so I think this is okay. I think
overall this is an improvement, as it merges in the levelWorker code
into our main bundle, and fixes the cross-origin error on production.

closes #22