Fix bug where controls are disabled after removing object
The containing element's focus events aren't that reliable for
determining focus: it's possible to permanently disable the controls
like this:
* Add object to the scene
* Remove object

This fixes that by explicitly enabling controls in this scenario.

Also, we don't need to touch the `controls2` instance - I'm not sure
what that's used for but we're only listening for key events on
`controls`.