Function Graph object: fix parse error in main function
This fixes a bug where the JS can crash by introducing a parse error
into the main function, e.g. mis-matched parentheses. Just delete the
final paren from the default expression and press enter, to reproduce
this bug.

These changes fix that, and allow the error to be displayed properly to
the user.