Fix bug where first "Make room" creates room as student
This is a quick fix to fix the bug that can be reproduced with these
steps:
1. open 3demos in private tab
2. click Session -> Make Room

User will be Student rather than Host.

Django doesn't create the user's session by default. We created this
explicitly in the room view, but not the index view, causing this bug.

There may be a nicer fix to enable this globally, but for now this
works.