Only create session when necessary
Minor adjustment for my fix for session creation: we currently only use
the session when the user wants to make a room. So, only create one on the
form POST.

This way, the fix is still in place, but session stuff won't be touched
at all if the user isn't touching the session functionality.